+++ src/components/fieldset/Fieldset.form.js
@@ -1,0 +1,4 @@
+import ComponentsEditForm from '../Components.form';
+export default function(...extend) {
+  return ComponentsEditForm(...extend);
+};